==> 01-Getting-Started.ipynb <==
---
Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk  
Spring 2024

---

# 1: Coding Basics

### Lesson Objectives
- Examine where Python code can be written and run
  - IPython & Jupyter Notebooks
  - Integrated Development Environments (IDEs): Spyder
- Review the basic elements of Python code:
  - Basic Python syntax: whitespace, comments, variable naming
  - Fundamental data types: numbers, strings, lists/tuples, sets, dictionaries
  - Python functions and methods
  - Getting help in Python

##  Tour of Jupyter Notebooks
Jupyter Notebooks are similar to R Markdown documents in that they include both formatted text and code. Jupyter Notebooks are quite powerful, but we haven't the time to fully explore them. Instead, I'll refer you to [Jupyter's documentation](https://jupyter.readthedocs.io/en/latest/index.html) to learn more and instead focus on what we need to know to use Jupyter to learn Python. 

### Creating your Jupyter environment
First, we'll be using Jupyter installed in one of Duke's "virtual containers". 
* Navigate to https://cmgr.oit.duke.edu/containers and log in.
* Find and click the button to (`reserve Jupyter`)[https://cmgr.oit.duke.edu/containers/Jupyter].

***→This adds the Jupyter container to the list of your reservations. This container will persist for the remainder of the semester and all your work will be saved on this virtual machine. You can return to it from any web browser using the link above!***

#### The Jupyter Console
The Jupyter Console is a sort of file manager. You'll see folders and some files with an `.ipynb` extension. Those files are Jupyter notebooks. Let's explore how these notebook files are used an managed. 

##### Opening an existing document
* Click on the `ggplot-python.ipynb` file. The notebook opens in a new tab. 
* Go back to the browser tab for the Jupyter console. Note that the `ggplot-python.ipynb` notebook is a different color. *This means that the kernel for that notebook is running. Even if you close the web page for that notebook, it remains "running" meaning all variables continue to be saved.* 
* Select the running notebook, then click the `Shutdown` button. 

##### Creating a new notebook
* From the Jupyter Console, select the `New ▼` dropdown, then select Python 3. 
* In the new document that appears, rename it `MyFirstNotebook`
* Save it
* Add a `print ("Hello World")` in the code block. 
* Run the code by clicking the `►|` button. 

**There, you've run a python command!** We'll learn more about Jupyter as we dig deeper into Python.

---
