==> 04-Data-Wrangling_II.ipynb <==
---
Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk  
Spring 2024

---
# 4: Data Wrangling II (in Python & Pandas)


## LESSON OBJECTIVES
1. Wrangle datasets with Python and Pandas functions
2. Compare R's `dplyr` to Python's `pandas` package with respect to wrangling data.
3. Apply data wrangling skills to a real-world example dataset

## SET UP YOUR DATA ANALYSIS SESSION
* Import Pandas
* Read in the following datasets in Pandas dataframes
 * `NTL-LTER_Lake_ChemistryPhysics_PeterPaul_Processed.csv`
 * `NTL-LTER_Lake_Nutrients_Raw.csv`

#Import pandas and numpy
import pandas as pd 
import numpy as np

#NTL.phys.data.PeterPaul <- read.csv("./Data/Processed/NTL-LTER_Lake_ChemistryPhysics_PeterPaul_Processed.csv")
NTL_physical = pd.read_csv('./data/Processed_KEY/NTL-LTER_Lake_ChemistryPhysics_PeterPaul_Processed.csv')

#NTL.nutrient.data <- read.csv("./Data/Raw/NTL-LTER_Lake_Nutrients_Raw.csv")
NTL_nutrient= pd.read_csv('./data/Raw/NTL-LTER_Lake_Nutrients_Raw.csv')

## REVIEW OF BASIC DATA EXPLORATION AND WRANGLING

# head(NTL.phys.data.PeterPaul)
NTL_physical.head()

#colnames(NTL.phys.data.PeterPaul)
NTL_physical.columns

#dim(NTL.phys.data.PeterPaul)
NTL_physical.shape

#summary(NTL.phys.data.PeterPaul$comments)
NTL_physical['comments'].value_counts(dropna=False)

#class(NTL.phys.data.PeterPaul$sampledate)
NTL_physical['sampledate'].dtype

* Format `sampledate` as date

#NTL.phys.data.PeterPaul$sampledate <- as.Date(NTL.phys.data.PeterPaul$sampledate, format = "%m/%d/%y")
NTL_physical['sampledate'] = pd.to_datetime(NTL_physical['sampledate'],format = "%Y/%m/%d")
NTL_physical['sampledate'].dtype

* Select Peter and Paul Lakes from the nutrient dataset

#NTL.nutrient.data.PeterPaul <- filter(NTL.nutrient.data, lakename == "Paul Lake" | lakename == "Peter Lake")
NTL_nutrient_PP = NTL_nutrient.query('lakename == "Paul Lake" | lakename == "Peter Lake"')

* Data summaries for nutrient data

#head(NTL.nutrient.data.PeterPaul)
NTL_nutrient_PP.head()

#colnames(NTL.nutrient.data.PeterPaul)
NTL_nutrient_PP.columns

#dim(NTL.nutrient.data.PeterPaul)
NTL_nutrient_PP.shape

#summary(NTL.nutrient.data.PeterPaul$comments)
NTL_nutrient_PP['comments'].value_counts(dropna=False)

#class(NTL.nutrient.data.PeterPaul$sampledate)
NTL_nutrient_PP['sampledate'].dtype

#NTL.nutrient.data.PeterPaul$sampledate <- as.Date(NTL.nutrient.data.PeterPaul$sampledate, format = "%m/%d/%y")
NTL_nutrient_PP['sampledate'] = pd.to_datetime(NTL_nutrient['sampledate'],format= "%m/%d/%y")

**_Oops! What's that warning??_**<br>
_The warning you get has to do with the fact that we are applying a calculation to a virtual **view** of a dataframe, i.e. a selection of records. `NTL_nutrient_PP` is only showing records corresponding to Paul Lake and Peter Lake, but it still knows it's linked to the full dataset. You can read more about this in the link provided in the error, or [this page](https://medium.com/dunder-data/selecting-subsets-of-data-in-pandas-part-4-c4216f84d388) has a better explanation of the isses._ 

_While this warning (not error) has no impact on our analysis, if you wanted to be super careful, an easy workaround is to create a "deep" copy of our subset dataframe. When we do this, it creates a new object, breaking the link..._ 

#Add "copy(deep=True)" to our query to ensure that our product is its own object, not a virtual view of another
NTL_nutrient_PP = NTL_nutrient.query('lakename == "Paul Lake" | lakename == "Peter Lake"').copy(deep=True)

#Now perform the date conversion. No error!
NTL_nutrient_PP['sampledate'] = pd.to_datetime(NTL_nutrient_PP['sampledate'],format= "%m/%d/%y")

NTL_nutrient_PP['sampledate'].dtype

* Save processed nutrient file

#write.csv(NTL.nutrient.data.PeterPaul, row.names = FALSE, file = "./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaul_Processed.csv")
NTL_nutrient_PP.to_csv("./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaul_Processed2.csv",index=False)

* Remove columns that are not of interest for analysis

#NTL.phys.data.PeterPaul.skinny <- select(NTL.phys.data.PeterPaul, 
#                                         lakename, daynum, year4, sampledate:irradianceDeck)
NTL_physical_skinny = NTL_physical[['lakename','daynum','year4','sampledate','depth', 'temperature_C',
                                    'dissolvedOxygen','irradianceWater','irradianceDeck']]

#NTL.nutrient.data.PeterPaul.skinny <- select(NTL.nutrient.data.PeterPaul, 
#                                             lakename, daynum, year4, sampledate, depth:po4)
NTL_nutrient_PP_skinny = NTL_nutrient_PP[['lakename','daynum','year4','sampledate','depth', 'tn_ug',
                                          'tp_ug', 'nh34', 'no23', 'po4']]

#NTL.phys.data.PeterPaul.skinny <- select(NTL.phys.data.PeterPaul, 
#                                         lakename, daynum, year4, sampledate:irradianceDeck)
NTL_physical_skinny = NTL_physical.loc[:,['lakename','daynum','year4','sampledate','depth', 'temperature_C',
                                          'dissolvedOxygen','irradianceWater','irradianceDeck']]

#NTL.nutrient.data.PeterPaul.skinny <- select(NTL.nutrient.data.PeterPaul, 
#                                             lakename, daynum, year4, sampledate, depth:po4)
NTL_nutrient_PP_skinny = NTL_nutrient_PP.loc[:,['lakename','daynum','year4','sampledate','depth', 'tn_ug',
                                                'tp_ug', 'nh34', 'no23', 'po4']]

#Alternatively...
NTL_phys_skinny = NTL_physical.drop(columns=['lakeid','comments'])
NTL_nutrient_PP_skinny = NTL_nutrient_PP.drop(columns=['lakeid','depth_id','comments'])

## TIDY DATASETS

For most situations, data analysis works best when you have organized your data into a tidy dataset. A tidy dataset is defined as: 

- Each variable is a column
- Each row is an observation (e.g., sampling event from a specific date and/or location)
- Each value is in its own cell

However, there may be situations where we want to reshape our dataset, for example if we want to facet numerical data points by measurement type (more on this in the data visualization unit). We can program this reshaping in a few short lines of code using the package tidyr, which is conveniently included in the tidyverse package. 

→ Help on tidying Pandas dataframes is [here](https://pandas.pydata.org/pandas-docs/stable/user_guide/reshaping.html)


* **Gather**, or **`melt`** in Pandas, nutrient data into one column

#Have a look at what we're going to melt
NTL_nutrient_PP_skinny.head()

#NTL.nutrient.data.PeterPaul.gathered <- 
#   gather(NTL.nutrient.data.PeterPaul.skinny, "nutrient", "concentration", tn_ug:po4)
NTL_nutrient_PP_gathered = NTL_nutrient_PP_skinny.melt(id_vars=['lakename','year4','daynum','sampledate','depth'],
                                                       var_name='nutrient',
                                                       value_name='concentration',
                                                       value_vars=['tn_ug','tp_ug','po4','nh34','no23','po4'])

→ _Unlike R's `gather`, Pandas' `melt` does not keep "placeholder `NA` values", i.e. the counts are different_

#count(NTL.nutrient.data.PeterPaul.gathered)
NTL_nutrient_PP_gathered['nutrient'].value_counts()

→ _Just to make sure, we'll drop NAs and reexamine the counts_

#NTL.nutrient.data.PeterPaul.gathered <- subset(NTL.nutrient.data.PeterPaul.gathered, !is.na(concentration))
NTL_nutrient_PP_gathered = NTL_nutrient_PP_gathered[~NTL_nutrient_PP_gathered['concentration'].isnull()]

#count(NTL.nutrient.data.PeterPaul.gathered, nutrient)
NTL_nutrient_PP_gathered['nutrient'].value_counts()

#write.csv(NTL.nutrient.data.PeterPaul.gathered, row.names = FALSE, 
#          file ="./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaulGathered_Processed.csv")
NTL_nutrient_PP_gathered.to_csv("./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaulGathered_Processed2.csv",
                                index=False
                               )

* **Spread (`pivot_table`)** nutrient data into separate columns

#NTL.nutrient.data.PeterPaul.spread <- spread(NTL.nutrient.data.PeterPaul.gathered, nutrient, concentration)
NTL_nutrient_PP_spread = NTL_nutrient_PP_gathered.pivot_table(columns='nutrient',
                                                              index=['lakename','year4','daynum','sampledate','depth'],
                                                              values='concentration').reset_index()
NTL_nutrient_PP_spread.head()

* **Split** components of cells into multiple columns. (Opposite of 'separate' is 'unite')

#NTL.nutrient.data.PeterPaul.dates <- separate(NTL.nutrient.data.PeterPaul.skinny, sampledate, c("Y", "m", "d"))
#NTL_nutrient_PP['dates'] = 
dtIndex = pd.DatetimeIndex(NTL_nutrient_PP_skinny['sampledate'])
NTL_nutrient_PP_dates = NTL_nutrient_PP_skinny.copy(deep='True')
NTL_nutrient_PP_dates['Y']=dtIndex.year
NTL_nutrient_PP_dates['m']=dtIndex.month
NTL_nutrient_PP_dates['d']=dtIndex.day

NTL_nutrient_PP_dates.head()

→ _The above is specific to splitting dates. However, we can use pandas' `str` function which allows us to apply any string function to the values in the column_

#Split the lakename column into two, using the space as separator
NTL_nutrient_PP_skinny['lakename'].str.split(' ',expand=True).head()

## JOINING MULTIPLE DATASETS
In many cases, we will want to combine datasets into one dataset. If all column names match, the data frames can be combined with the [`concat`](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.concat.html) function. If some column names match and some column names don't match, we can combine the data frames using a **"merge"** function according to common conditions that exist in the matching columns. We will demonstrate this with the NTL-LTER physical and nutrient datasets, where we have specific instances when physical and nutrient data were collected on the same date, at the same lake, and at the same depth. 

In Pandas, there are several types of join functions: 

* `left`: use only keys from left frame, similar to a SQL left outer join; preserve key order.
* `right`: use only keys from right frame, similar to a SQL right outer join; preserve key order.
* `outer`: use union of keys from both frames, similar to a SQL full outer join; sort keys lexicographically.
* `inner`: use intersection of keys from both frames, similar to a SQL inner join; preserve the order of the left keys.

Let's say we want to generate a new dataset that contains all possible physical and chemical data for Peter and Paul Lakes. In this case, we want to do a full, or "outer" join, joining the tables on the fields that are common: `lakename`,`year4`,`daynum`,`sampledate`,and `depth`.

#First, examine the sizes of both dataframes
NTL_physical_skinny.head(2)

NTL_nutrient_PP_skinny.head(2)

#NTL.phys.nutrient.data.PeterPaul <- full_join(NTL.phys.data.PeterPaul.skinny, NTL.nutrient.data.PeterPaul.skinny) 
NTL_phys_nut_PP = pd.merge(left=NTL_physical_skinny,
                          right=NTL_nutrient_PP_skinny,
                          how='outer',
                          on = ["lakename","year4","daynum","sampledate","depth"]
                         )
#How many records in the joined result? 
NTL_phys_nut_PP.shape

NTL_phys_nut_PP.head()

---
<font color='blue'>► How many records are in the larger of the two unjoined dataframes. Would you expect an outer join to have more, fewer, or the same number of records as this dataframes? Why? (Check your answer in the cells below...)

#How many records in NTL_physical_skinny?
NTL_physical_skinny.shape

#How many records in NTL_nutrient_PP_skinny?
NTL_nutrient_PP_skinny.shape

<font color='blue'>► Change the join type to `inner`. Now, how to the sizes compare? What about `left`? `right`?</font>
<hr>

#write.csv(NTL.phys.nutrient.data.PeterPaul, row.names = FALSE, 
#          file ="./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv")
NTL_phys_nut_PP.to_csv('./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed2.csv',
                       index=False)

## GROUPING AND SUMMARIZING DATA (SPLIT-APPLY-COMBINE)

dplyr functionality, combined with the pipes operator, allows us to split datasets according to groupings (function: `group_by`), then run operations on those groupings and return the output of those operations. There is a lot of flexibility in this approach, but we will illustrate just one example today.

The R code we'll mimick:
```{r}
NTL.PeterPaul.summaries <- 
  NTL.phys.nutrient.data.PeterPaul %>%
  filter(depth == 0) %>%
  group_by(lakename) %>%
  filter(!is.na(temperature_C) & !is.na(tn_ug) & !is.na(tp_ug)) %>%
  summarise(meantemp = mean(temperature_C),  #Can run any funcion here
            counts = n(),
            sdtemp = sd(temperature_C), 
            meanTN = mean(tn_ug), 
            sdTN = sd(tn_ug), 
            meanTP = mean(tp_ug), 
            sdTP = sd(tp_ug))


```

NTL_PP_Summary = (
    NTL_phys_nut_PP.
    dropna(how='any',axis='rows',subset=['temperature_C','tn_ug','tn_ug']).  #Filter rows with no NA in these fields
    query('depth == 0').                 #Select records where depth is 0
    groupby('lakename').                 #Group by the `lakename` field
    agg({'lakename':['count'],           #Compute aggregate count of outputs
         'temperature_C':['mean','std'], #Compute aggregate stats of `temperature_C` values by group
         'tn_ug':['mean','std'],         #Compute aggregate stats of `tn_ug` values by group
         'tp_ug':['mean','std']          #Compute aggregate stats of `tp_ug` values by group
        })
)       
NTL_PP_Summary

#write.csv(NTL.PeterPaul.summaries, row.names = FALSE, 
#          file ="./Data/Processed/NTL-LTER_Lake_Summaries_PeterPaul_Processed.csv")
NTL_PP_Summary.to_csv("./Data/Processed/NTL-LTER_Lake_Summaries_PeterPaul_Processed2.csv",index=False)
